requirements.txt
Flask
redis
python-dotenv
groq
sentence-transformers
numpy
gunicorn 